# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vireolin API",
        "description": "API توثيق لتطبيق Vireolin",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/serviceTypes": {
            "get": {
                "tags": [
                    "serviceTypes"
                ],
                "summary": "Get all service types",
                "operationId": "a40ad509983276d19b0c27f65c0faf5f",
                "responses": {
                    "200": {
                        "description": "List of service types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "code": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "serviceTypes"
                ],
                "summary": "Create service type",
                "operationId": "cd2e602aa38ae205e9977ccc86a46118",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Cleaning"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Cleaning service description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service type created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "code": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/serviceTypes/{id}": {
            "get": {
                "tags": [
                    "serviceTypes"
                ],
                "summary": "Get a single service type",
                "operationId": "58d8d55d9e8c5098dd21b704b0102a70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ServiceType ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Single service type info"
                    },
                    "404": {
                        "description": "Service type not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "serviceTypes"
                ],
                "summary": "Update service type",
                "operationId": "0181b1964257bbb5ac23321f1668a248",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ServiceType ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Service type updated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "serviceTypes"
                ],
                "summary": "Delete a service type",
                "operationId": "44fe20b3dd2e9f78a871feea03edeb66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ServiceType ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service type deleted"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}